requirements.txt
pandas>=2.2
pyarrow>=16.0
duckdb>=1.0
requests>=2.32
tqdm>=4.66
python-dotenv>=1.0
